requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.3.0
av==14.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
coloredlogs==15.0.1
dotenv==0.9.9
fast-alpr==0.1.2
fast-plate-ocr==0.3.0
flatbuffers==25.2.10
frozenlist==1.5.0
humanfriendly==10.0
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
libarchive-c==5.2
markdown-it-py==3.0.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
numpy==1.26.4
onnxruntime==1.19.2
open-image-models==0.2.1
opencv-python==4.11.0.86
packaging==24.2
paho-mqtt==2.1.0
picamera2==0.3.25
pidng==4.0.9
piexif==1.1.3
pillow==11.1.0
propcache==0.3.0
protobuf==6.30.0
Pygments==2.19.1
python-dotenv==1.0.1
python-prctl==1.8.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
RPi.GPIO==0.7.1
schedule==1.2.2
simplejpeg==1.8.2
sympy==1.13.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
v4l2-python3==0.3.5
yarl==1.18.3
